Added Real-time search and load model(#311)
At the beginning, there is some default mode showed as Search Result. Then, whatever the user input, the system will match on real time and display the related result. In the end, the user can choose the model from the options and click them to load directly. Therefore, the users don't need to find from the list, especially when the number of models becomes larger in the future. Besides, the users don't need to even remember the whole name of the model.